requirements.txt
speechrecognition
pyaudio
pyttsx3
google-generativeai
python-dotenv